Preventing infinite loop bug

+++ daemon/daemon.py
@@ -113,8 +113,9 @@ def execute_action(xmms_controller, command):
             elif split_command[0] == "delete":
                 song_number = int(split_command[2])
                 try:
-                    song = Song.objects.get(position=song_number)
-                    song.delete()
+                    if song_number:
+                        song = Song.objects.get(position=song_number)
+                        song.delete()
                     xmms_controller.delete(song_number)
                 except Exception:
                     pass
